refactor: simplify semantic diff report rendering
